refactor(mitm-addon): extract connector diagnostic owner
